Draw a tombstone in curses
Pretty code it is not. But it looks freaking rad.

"Write comically straightforward code, as if you just learned to
program, and go out of your way avoid wearing any kind of software
engineering hat--unless what you really want to be is a software
engineer, and not the designer of an experience." - Prog21