Use loaded state when starting game from file
Doi! This code was based on the assumption that these three values
were the only important ones in the State struct. After adding the
player data to State, this was no longer the case. So ditch the old
struct and use the loaded one instead.

Thanks to Disagi on Discord for catching this one.